[loop-arc] Be more explicit in the docs on how we handle loops with multiple backedges.
Previously there was just a flippant comment. Now it is clearly
specified that the head/tails of the back edges of multiple back edge
loop are marked with the same flag as we use to mark the head/tail of
irreducible loops. The only difference is that in the irreducible case,
we mark all blocks (irregardless of the blocks position as a head/tail)
as being part of an unknown control flow sequence. In the case of a
multiple back edge loop, we mark the heads of the back edges as unknown
control flow edge heads and perform the analogous operation for the
tails.

Swift SVN r32641