[loop-arc] Use FunctionTy instead of SILFunction.
I would like to keep this data structure as general as possible in case
I would like to use it in LLVM in the future. So I am trying to make
sure that I refer to typedefed types instead of types directly. This
will make it trivial to port.

Swift SVN r32642