Use SmallSetVector for callee sets.
SmallSetVector takes more space than SmallPtrSet, but the number of
these that we actually allocate is bounded by the number of unique
method decls we see in vtables, so total memory usage won't grow by
much.

This allows us to remove the sorting that we're doing at various points
in processing as well as enabling the use of ArrayRefView (in a later
commit) to hide call graph nodes and instead just vend the called
SILFunctions.

Swift SVN r32723