[loop-arc] Create canonicalizeAllLoops(DominanceInfo *, SILLoopInfo *LI)
This just canonicalizes all loops by going bottom up through the loop
nest and returns true if any loops were canonicalized. It is meant to be
used by loop arc to make sure that all loops are canonicalized before we
attempt to perform dataflow.

One thing that we should think about is potentially having a phase after
which it is a verification error for loops to no longer be canonical.

rdar://22238658

Swift SVN r32750