[loop-arc] Create ARC infrastructure needed to plug in loop ARC into the current ARC framework.
The second stage of bringing up loop ARC is to run normal ARC in a loop
ARC context. This should result in exactly the same output semantically
allowing us to have a "correct" baseline to work against as we add in
loop arc specific features.

This code is wired up in a subsequent commit.

rdar://22238658

Swift SVN r32756