First version of escape analysis.
The escape analysis checks if a pointer (e.g. an object reference) escapes the scope of a function,
for example because it is stored to a global variable.
For details see the comments in EscapeAnalysis.h

With this commit the escape analysis is computed (and verified) but not yet used in any optimization pass.
The plan is to use it in AllocBoxToStack and in a more general stack promotion.



Swift SVN r32761