[IRGen] Fix source list computation in PolymorphicConvention
Currently, a source for polymorphic parameter fulfillments is removed
from the source list if the size of the map of fulfillments does not
increase during its consideration. This isn't quite correct because it's
possible for a new fulfillment that uses the source to replace an
existing one without creating a new entry in the map. Have the code be
more explicit about the condition under which a source may be safely
discarded.

<rdar://problem/23121786> Crash emitting IR SIL function for Carthage

Swift SVN r32763