[loop-arc] Change RefCountState to be a non-CRTP construct and refactor code to subclasses where necessary.
The reason why this is necessary is that the interfaces for bottom up/top down
ref count states is going to diverge slightly so that I can pass in insertpts in
the appropriate places. If I am going to leak through abstraction I might as
well just remove the abstraction.

The reason why I need to explicitly pass in insertpts is that in the case of a
loop region, I want to be able to process instructions in the loop for effects,
but have the insert point be outside the loop.

This is a pure refactor commit that is meant to be as simple as possible and
just move code around. In a subsequent commit I will simplify the code to remove
the unnecessary CRTP related code.

rdar://22238658

Swift SVN r32918