[loop-arc] Pass in the insertion point into BottomUpRefCountState instead of computing it inside.
This will enable in the case of loops to use an insertion point outside the loop
even though the instruction that we are "applying" to the ref count is inside
the loop. I am going to do a similar commit in a little bit that does the same
thing for TopDownRefCountState.

Swift SVN r32922