Update edge removal API in call graph.
Add a parameter for a function the call site is in, rather than using
the call site to determine the function.

This is necessary for enabling invalidation of call graph information
for individual functions. If we're not maintaining the call graph during
a pass, we may delete an instruction and be left with a dangling pointer
in the call graph edge. This allows us to still remove the edge properly
without trying to look at the instruction it represents.