[loop-arc] Add a new ARC Analysis API called canNeverDecrementRefCounts(SILInstruction *)
For the first version of Loop ARC I am going to take a very conservative
approach that mostly relies on known safe for performance. The basic
idea is to just collect all instructions that may use or decrement a
value and then when processing an outside loop, apply the current
primitives to pointers to determine if anything inside of the loop could
decrement/use the values. If anything can decrement, we treat it as a
guaranteed use. If anything can just use the value, we treat it as a
use.

This API is needed for this purpose.

rdar://22238729