Use SILFileLocations temporarily on inserted retain, releases.
The current model of using insertion points for a retain/release's debug
location does not work in the case where the insertion point is a terminator
block.

This is necessary since sometimes ARC will need to insert a retain, release in a
block that only has a terminator. This means that the insert point will be a
terminator. Thus the invariant that return locations can only be on return insts
will be violated.

I filed rdar://23353815 to track the debug locations being re-emitted.