[loop-arc] Move ARCOpts specific analysis code into SILPasses from SILAnalysis.
This decision to split ARC in between SILAnalysis and SILPasses was originally
intended to allow for ARC to provide pairings of retain/releases for other
passes besides ARCOpts.

For a while we have known that the current form of the ARC optimizer is not
amenable for this b/c it does not do semantic pairings.

When we actually perform real semantic pairing analysis, I will do it in
SILAnalysis.

I left the actual generic ARC code in SILAnalysis.