Do not perform any retain sinking in the speculative devirtualizer.
First of all, trying to optimize any RC-operations is not really a task of the devirtualizer. And then, the logic used to do it was too naive anyways.
We have more powerful dedicated optimization passes now (e.g. SIL code motion), which can do the same kind of retain sinking and even more.