IRGen: optionals of blocks now use block reference counting entry points
Replace isSingle{Unknown,Swift}ReferenceCountedObject() with a single
entry point that also returns the reference counting style. Use this
in GenEnum to emit more specific entry points than native and unknown.

This will give a slight performance boost on Darwin, and enable use of
the blocks runtime on Linux.

Progress on <rdar://problem/23315750>.