Debug Info: Establish a default ordering for compile units.
Because the swift compiler relies on Clang to setup the Module,
the clang CU is always created first.  Several dwarf-reading
tools (older versions of ld64, and lldb) can get confused if the
first CU in an object is empty, so ensure that the Swift CU comes
first by rearranging the list of CUs in the LLVM module.

rdar://problem/23434643