Remove code sinking logic from the speculative devirtualizer.
The code sinking pass (that runs after the speculative devirtualize) can already
sink instructions.  The decision to sink instructions is non trivial if the
instructions has side effects and we'd like to keep this logic in one place.

Roman recently removed the code for removing retains and this commit removes
the code for sinking class_method instructions.