Add support for devirtualizing init/deinit.
Due to the way a helper function was written we previously did not
handle these. This also makes it possible for the call graph code to
properly handle these (although the remaining support for deinits is
coming in a future commit).

You might notice that when we devirtualize try_apply we generate worse
code than we might otherwise. This was the case before my change, but my
change exposes it in another test case. I opened rdar://problem/23458358
to track improvements to this.