[Inliner] Migrate the cycle-prevention mechanism from pointers to strings.
We are working on moving away from the bump-pointer allocator to malloc/free.
The inliner keeps a list of pointers to deleted instructions. When we move to
malloc/free scheme instructions can be re-allocated on older pointers and break
this system.

To work around this problem this commit moves to a different approach. Instead
of saving a list of pointers we are keeping a list of function names.  After
each iteration of inlining functions into the caller function we merge the local
log to a global list.  On our next inlining iteration we check the log to see if
the function was already inlined.

Along the way fix rdar://22936260 (endless inlining loop).