Remove use of CallGraph in Function Signature Optimization.
The CallGraph Analysis (as we know it today) is going away. This commit replaces
the use of the call graph with a local implementation of a basic scan of the
module that constructs the caller-callee listing. The only functionality that we
keep is the bottom-up function order, and this is not going away.