[stdlib] Integer prototype: remove promotion model
A promotion model is not really implementable in a library with the
language as it stands today.  Protocols can express a function from one
type to another, but without some way to express a function from two
types to a result type, any implementation of promotion is a hack that
at best stresses the type checker and doesn't really work generically.

We can circle back and implement heterogeneous *in-place*
arithmetic (e.g. +=) once the algorithms needed by BigInt are
implemented.

This reverts commits 13079577d9d68eb3e8d5b607525e034f80969299 and
23ac13751152333f674739d7ed5274d63ea7ea3d.