Fix use-after-free error caused by SILBuilderWithScope.
The problem occurs if one uses a SILBuilderWithScope to create an instruction
and then later on delete that instruction before the SILBuilderWithScope's
destructor has been called. In such a case when the SILBuilderWithScope's
destructor runs, the destructor will attempt to set the debug scope of the
deleted instruction.

This will be caught by ASAN once a patch by Roman lands that changes
instructions to use malloc instead of BumpPtrAllocators.

<rdar://problem/23548378>